Use UTC for milestone proposedCompletionDate
The proposedCompletionDate field is saved as a date on the database (only the date), but when drizzle get the data it's converted to a date with the time set to 00:00:00 with UTC as the timezone. So then, when you try to show it, it's converted to the local GMT and the date can be shown wrong.